fix: moving methods around
